optimized

+++ src/facade/NewOrder.java
@@ -1,7 +1,7 @@
 package facade;
 
 public class NewOrder {
-    public void createOrder(String product,double price){
+    public void createOrder(String product,float price){
         System.out.println("Order created: " + product+" "+price+"tenge");
     }
 }
