+++ src/facade/Payment.java
@@ -1,9 +1,7 @@
 package facade;
 
 public class Payment {
-
-    public void processPayment(double sum) {
+    public void processPayment(float sum) {
         System.out.println("Payment processed: " + sum+" tenge");
     }
-
 }
